feat: upgrade OpenAudio backend to Fish Speech S2-Pro and refactor dependencies
