+++ inference/src/backends/registry.py
@@ -1,7 +1,7 @@
 from .base import TTSBackend
 from .chatterbox import ChatterboxBackend
 from .cosyvoice import CosyVoiceBackend
-from .openaudio import OpenAudioBackend
+from .fish_audio import FishAudioBackend
 from .kokoro import KokoroBackend
 from .qwen import QwenBackend
 from .f5tts import F5TTSBackend
@@ -14,7 +14,7 @@ _BACKENDS: dict[str, type[TTSBackend]] = {
     "f5tts": F5TTSBackend,
     "piper": PiperBackend,
     "cosyvoice": CosyVoiceBackend,
-    "openaudio": OpenAudioBackend,
+    "fish_audio": FishAudioBackend,
     "chatterbox": ChatterboxBackend,
     "higgs_audio": HiggsAudioBackend,
 }
