+++ server/src/manifest/models.manifest.ts
@@ -207,20 +207,28 @@ export const modelsCatalog: CatalogModel[] = [
     maxReferenceDuration: 20,
     supportsInstruct: true,
   },
-  // ── OpenAudio S1-mini (multilingual cloning) ───────────────────────
+  // ── Fish Speech S2-Pro (multilingual cloning with emotion control) ──
   {
-    id: 'openaudio-s1-mini',
-    name: 'OpenAudio S1 Mini',
-    backend: 'openaudio',
-    backendDisplayName: 'OpenAudio',
-    backendDescription: 'High-quality multilingual voice cloning with emotional expression by Fish Audio.',
-    description: 'OpenAudio S1-mini by Fish Audio — high-quality multilingual voice cloning with emotional expression.',
-    repo: 'fishaudio/openaudio-s1-mini',
-    files: ['config.json', 'model.pth', 'codec.pth', 'special_tokens.json', 'tokenizer.tiktoken'],
-    size: 3_610_000_000,
+    id: 'fish-speech-s2-pro',
+    name: 'Fish Speech S2-Pro',
+    backend: 'fish_audio',
+    backendDisplayName: 'Fish Speech',
+    backendDescription: 'SOTA multilingual voice cloning with inline emotion control by Fish Audio (4B Dual-AR).',
+    description: 'Fish Speech S2-Pro by Fish Audio — SOTA multilingual TTS with zero-shot voice cloning and inline emotion/prosody control (4B Dual-AR, 80+ languages).',
+    repo: 'fishaudio/s2-pro',
+    files: [
+      'config.json',
+      'codec.pth',
+      'model.safetensors.index.json',
+      'model-00001-of-00002.safetensors',
+      'model-00002-of-00002.safetensors',
+      'special_tokens_map.json',
+      'tokenizer.json',
+      'tokenizer_config.json',
+    ],
+    size: 9_000_000_000,
     types: ['cloning'],
     maxReferenceDuration: 25,
-    gated: true,
     supportsEffects: true,
   },
   // ── Chatterbox (multilingual cloning with emotion control) ──────────
